+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This release is a big one with the introduction of a `Wgit::DSL` and Javascript 
 - `Wgit::Utils.sanitize` which calls `.sanitize_*` underneath.
 - `Wgit::Crawler#crawl_site` now has a `follow:` named param - if set, it's xpath value is used to retrieve the next urls to crawl. Otherwise the `:default` is used (as it was before). Use this to override how the site is crawled.
 - `Wgit::Database` methods: `#clear_urls`, `#clear_docs`, `#clear_db`, `#text_index`, `#text_index=`, `#create_collections`, `#create_unique_indexes`, `#docs`, `#get`.
+- `Wgit::Database#clear_db!` alias.
 - `Wgit::Document` methods: `#at_xpath`, `#at_css` - which call nokogiri underneath.
 - `Wgit::Indexer#index_urls` method which can index several urls in one call.
 - `Wgit::Document#extract` method to perform one off content extractions.